+++ sdk/src/driftClient.ts
@@ -102,6 +102,7 @@ import {
 	getUserAccountPublicKey,
 	getUserAccountPublicKeySync,
 	getUserStatsAccountPublicKey,
+	getSignedMsgWsDelegatesAccountPublicKey,
 } from './addresses/pda';
 import {
 	DataAndSlot,
@@ -1156,6 +1157,108 @@ export class DriftClient {
 		return resizeUserAccountIx;
 	}
 
+	public async initializeSignedMsgWsDelegatesAccount(
+		authority: PublicKey,
+		delegates: PublicKey[] = [],
+		txParams?: TxParams
+	): Promise<TransactionSignature> {
+		const ix = await this.getInitializeSignedMsgWsDelegatesAccountIx(
+			authority,
+			delegates
+		);
+		const tx = await this.buildTransaction([ix], txParams);
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+		return txSig;
+	}
+
+	public async getInitializeSignedMsgWsDelegatesAccountIx(
+		authority: PublicKey,
+		delegates: PublicKey[] = []
+	): Promise<TransactionInstruction> {
+		const signedMsgWsDelegates = getSignedMsgWsDelegatesAccountPublicKey(
+			this.program.programId,
+			authority
+		);
+		const ix = await this.program.instruction.initializeSignedMsgWsDelegates(
+			delegates,
+			{
+				accounts: {
+					signedMsgWsDelegates,
+					authority: this.wallet.publicKey,
+					rent: anchor.web3.SYSVAR_RENT_PUBKEY,
+					systemProgram: anchor.web3.SystemProgram.programId,
+				},
+			}
+		);
+		return ix;
+	}
+
+	public async addSignedMsgWsDelegate(
+		authority: PublicKey,
+		delegate: PublicKey,
+		txParams?: TxParams
+	): Promise<TransactionSignature> {
+		const ix = await this.getAddSignedMsgWsDelegateIx(authority, delegate);
+		const tx = await this.buildTransaction([ix], txParams);
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+		return txSig;
+	}
+
+	public async getAddSignedMsgWsDelegateIx(
+		authority: PublicKey,
+		delegate: PublicKey
+	): Promise<TransactionInstruction> {
+		const signedMsgWsDelegates = getSignedMsgWsDelegatesAccountPublicKey(
+			this.program.programId,
+			authority
+		);
+		const ix = await this.program.instruction.changeSignedMsgWsDelegateStatus(
+			delegate,
+			true,
+			{
+				accounts: {
+					signedMsgWsDelegates,
+					authority: this.wallet.publicKey,
+					systemProgram: anchor.web3.SystemProgram.programId,
+				},
+			}
+		);
+		return ix;
+	}
+
+	public async removeSignedMsgWsDelegate(
+		authority: PublicKey,
+		delegate: PublicKey,
+		txParams?: TxParams
+	): Promise<TransactionSignature> {
+		const ix = await this.getRemoveSignedMsgWsDelegateIx(authority, delegate);
+		const tx = await this.buildTransaction([ix], txParams);
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+		return txSig;
+	}
+
+	public async getRemoveSignedMsgWsDelegateIx(
+		authority: PublicKey,
+		delegate: PublicKey
+	): Promise<TransactionInstruction> {
+		const signedMsgWsDelegates = getSignedMsgWsDelegatesAccountPublicKey(
+			this.program.programId,
+			authority
+		);
+		const ix = await this.program.instruction.changeSignedMsgWsDelegateStatus(
+			delegate,
+			false,
+			{
+				accounts: {
+					signedMsgWsDelegates,
+					authority: this.wallet.publicKey,
+					systemProgram: anchor.web3.SystemProgram.programId,
+				},
+			}
+		);
+		return ix;
+	}
+
 	public async initializeFuelOverflow(
 		authority?: PublicKey
 	): Promise<TransactionSignature> {
@@ -2072,7 +2175,11 @@ export class DriftClient {
 			this.getRemainingAccountMapsForUsers(params.userAccounts);
 
 		if (params.useMarketLastSlotCache) {
-			const lastUserSlot = this.getUserAccountAndSlot()?.slot;
+			const lastUserSlot = this.getUserAccountAndSlot(
+				params.userAccounts.length > 0
+					? params.userAccounts[0].subAccountId
+					: this.activeSubAccountId
+			)?.slot;
 
 			for (const [
 				marketIndex,
@@ -2480,7 +2587,7 @@ export class DriftClient {
 		let remainingAccounts = [];
 		if (userInitialized) {
 			remainingAccounts = this.getRemainingAccounts({
-				userAccounts: [await this.forceGetUserAccount()],
+				userAccounts: [await this.forceGetUserAccount(subAccountId)],
 				useMarketLastSlotCache: true,
 				writableSpotMarketIndexes: [marketIndex],
 			});
@@ -3322,6 +3429,71 @@ export class DriftClient {
 					borrowToSpotMarketVault:
 						this.getSpotMarketAccount(borrowToMarketIndex).vault,
 					driftSigner: this.getSignerPublicKey(),
+				},
+				remainingAccounts,
+			}
+		);
+	}
+
+	public async transferPerpPosition(
+		fromSubAccountId: number,
+		toSubAccountId: number,
+		marketIndex: number,
+		amount: BN,
+		txParams?: TxParams
+	): Promise<TransactionSignature> {
+		const { txSig } = await this.sendTransaction(
+			await this.buildTransaction(
+				await this.getTransferPerpPositionIx(
+					fromSubAccountId,
+					toSubAccountId,
+					marketIndex,
+					amount
+				),
+				txParams
+			),
+			[],
+			this.opts
+		);
+		return txSig;
+	}
+
+	public async getTransferPerpPositionIx(
+		fromSubAccountId: number,
+		toSubAccountId: number,
+		marketIndex: number,
+		amount: BN
+	): Promise<TransactionInstruction> {
+		const fromUser = await getUserAccountPublicKey(
+			this.program.programId,
+			this.wallet.publicKey,
+			fromSubAccountId
+		);
+		const toUser = await getUserAccountPublicKey(
+			this.program.programId,
+			this.wallet.publicKey,
+			toSubAccountId
+		);
+
+		const remainingAccounts = this.getRemainingAccounts({
+			userAccounts: [
+				this.getUserAccount(fromSubAccountId),
+				this.getUserAccount(toSubAccountId),
+			],
+			useMarketLastSlotCache: true,
+			writablePerpMarketIndexes: [marketIndex],
+		});
+
+		return await this.program.instruction.transferPerpPosition(
+			marketIndex,
+			amount ?? null,
+			{
+				accounts: {
+					authority: this.wallet.publicKey,
+					fromUser,
+					toUser,
+					userStats: this.getUserStatsAccountPublicKey(),
+					state: await this.getStatePublicKey(),
 				},
 				remainingAccounts,
 			}
@@ -9121,6 +9293,37 @@ export class DriftClient {
 		return [verifyIx, ix];
 	}
 
+	public async getPostManySwitchboardOnDemandUpdatesAtomicIxs(
+		feeds: PublicKey[],
+		recentSlothash?: Slothash,
+		numSignatures = 3
+	): Promise<TransactionInstruction[] | undefined> {
+		const program = await this.getSwitchboardOnDemandProgram();
+		for (const feed of feeds) {
+			const feedAccount = new PullFeed(program, feed);
+			if (!this.sbProgramFeedConfigs) {
+				this.sbProgramFeedConfigs = new Map();
+			}
+			if (!this.sbProgramFeedConfigs.has(feedAccount.pubkey.toString())) {
+				const feedConfig = await feedAccount.loadConfigs();
+				this.sbProgramFeedConfigs.set(feed.toString(), feedConfig);
+			}
+		}
+
+		const [pullIxs, _responses, success] =
+			await PullFeed.fetchUpdateManyLightIx(program, {
+				feeds,
+				numSignatures,
+				recentSlothashes: recentSlothash
+					? [[new BN(recentSlothash.slot), recentSlothash.hash]]
+					: undefined,
+			});
+		if (!success) {
+			return undefined;
+		}
+		return pullIxs;
+	}
+
 	public async getPostSwitchboardOnDemandUpdateAtomicIx(
 		feed: PublicKey,
 		recentSlothash?: Slothash,
@@ -9135,18 +9338,20 @@ export class DriftClient {
 			const feedConfig = await feedAccount.loadConfigs();
 			this.sbProgramFeedConfigs.set(feed.toString(), feedConfig);
 		}
-		const [pullIx, _responses, success] = await feedAccount.fetchUpdateIx(
+		const [pullIx, _responses, success] = await PullFeed.fetchUpdateManyIx(
+			program,
 			{
+				feeds: [feed],
 				numSignatures,
-			},
-			recentSlothash
-				? [[new BN(recentSlothash.slot), recentSlothash.hash]]
-				: undefined
+				recentSlothashes: recentSlothash
+					? [[new BN(recentSlothash.slot), recentSlothash.hash]]
+					: undefined,
+			}
 		);
 		if (!success) {
 			return undefined;
 		}
-		return pullIx;
+		return pullIx[0];
 	}
 
 	public async postSwitchboardOnDemandUpdate(