+++ CHANGELOG.md
@@ -13,6 +13,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking
 
+## [2.114.0] - 2025-03-13
+
+### Features
+
+### Fixes
+
+- program: add liq fees to calculate_perp_market_amm_summary_stats ([#1531](https://github.com/drift-labs/protocol-v2/pull/1531))
+
+### Breaking
+
+## [2.113.0] - 2025-03-06
+
+### Features
+
+- program: add transfer_perp_position ix ([#1514](https://github.com/drift-labs/protocol-v2/pull/1514))
+- program: add signed_msg ws delegate account ([#1515](https://github.com/drift-labs/protocol-v2/pull/1515))
+
+### Fixes
+
+### Breaking
+
 ## [2.112.0] - 2025-03-03
 
 ### Features